feat: business 공통 tab 적용   ( #102)
