+++ frontend/src/pages/BProfile/style.jsx
@@ -1,0 +1,4 @@
+import styled from "styled-components";
+
+export const ProfileTab1 = styled.div``;
+export const ProfileTab2 = styled.div``;
